more changes and remove off old file (renamde)
